requirements.txt
# File location: /requirements.txt
# S.H.I.T. Bot Python Dependencies

# Core
python-telegram-bot==21.0.1
python-dotenv==1.0.0

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9
aiosqlite==0.19.0

# Web3 & Blockchain
web3==6.11.3
eth-account==0.10.0
eth-utils==2.3.1

# Data Processing
pandas==2.1.4
openpyxl==3.1.2
xlsxwriter==3.1.9

# HTTP & Async
aiohttp==3.9.1
httpx==0.25.2
asyncio==3.4.3

# Utilities
python-dateutil==2.8.2
pytz==2023.3
Pillow==10.1.0

# Ollama Integration
ollama==0.1.7

# Logging & Monitoring
colorlog==6.8.0
prometheus-client==0.19.0

# Calendar Widget
python-telegram-calendar-package==1.0.5

# JSON & Data Validation
jsonschema==4.20.0
pydantic==2.5.2

# Development & Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.12.0
flake8==6.1.0
mypy==1.7.1

# Rate Limiting
aioredis==2.0.1
aiocache==0.12.2

# Performance
ujson==5.9.0
orjson==3.9.10

# Security
cryptography==41.0.7
python-jose==3.3.0